Fixing d-pointer
